LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/perspective-origin-expected.txt

CONSOLE MESSAGE: Error: '<position>' is not a valid CSS component

Harness Error (FAIL), message = Error: '<position>' is not a valid CSS component

FAIL Can set 'perspective-origin' to CSS-wide keywords assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'perspective-origin' to var() references assert_equals: expected 2 but got 1
FAIL Can set 'perspective-origin' to the 'none' keyword Bad value for shorthand CSS property
